quick-draw/requirements.txt
apache-beam==2.6.0
avro==1.8.2
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
crcmod==1.7
dill==0.2.8.2
docopt==0.6.2
enum34==1.1.6
fastavro==0.19.7
fasteners==0.14.1
funcsigs==1.0.2
future==0.16.0
futures==3.2.0
gapic-google-cloud-pubsub-v1==0.15.4
google-apitools==0.5.20
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-cloud-bigquery==0.25.0
google-cloud-core==0.25.0
google-cloud-pubsub==0.26.0
google-gax==0.15.16
googleapis-common-protos==1.5.3
googledatastore==7.0.1
grpc-google-iam-v1==0.11.4
grpcio==1.14.2
hdfs==2.1.0
httplib2==0.11.3
idna==2.7
mock==2.0.0
monotonic==1.5
numpy==1.15.1
oauth2client==4.1.2
pbr==4.2.0
Pillow==5.2.0
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-pubsub-v1==0.15.4
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pydot==1.2.4
pyparsing==2.2.0
pytz==2018.4
PyVCF==0.6.8
PyYAML==3.13
requests==2.19.1
rsa==3.4.2
six==1.11.0
typing==3.6.6
urllib3==1.23
